refactor: melhorias no método de deletar usuarios

+++ cadastro-usuario/src/main/java/com/luanmiranda/cadastro_usuario/business/UsuarioService.java
@@ -34,7 +34,16 @@ public class UsuarioService {
         );
     }
 
-    public void excluirUsuario(String email) {
-        usuarioRepository.deleteByEmail(email);
+    public boolean excluirUsuario(String email) {
+        try {
+            if (usuarioRepository.existsByEmail(email)) {
+                usuarioRepository.deleteByEmail(email);
+                return true;
+            } else {
+                return false;
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("Erro ao excluir usuário", e);
+        }
     }
 }
